% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-frame.R
\name{check_frame}
\alias{check_frame}
\title{Run a list of checks against a data frame}
\usage{
check_frame(data, checks)
}
\arguments{
\item{data}{A data frame.}

\item{checks}{A list of check functions, e.g. created by \code{\link[=chk_has_columns]{chk_has_columns()}}.}
}
\value{
An object of class \code{framecheck_report}.
}
\description{
Run a list of checks against a data frame
}
\examples{
df <- data.frame(id = c(1, 2), score = c(5, 99))
check_frame(df, list(
  chk_has_columns(c("id", "score")),
  chk_unique_key("id"),
  chk_in_range("score", 0, 10)
))
}
