% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{chk_col_type}
\alias{chk_col_type}
\title{Check that a column has the expected class}
\usage{
chk_col_type(col, type)
}
\arguments{
\item{col}{Column name.}

\item{type}{Expected first class, e.g. "numeric".}
}
\value{
A check function taking a data frame.
}
\description{
Check that a column has the expected class
}
