% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{chk_unique_key}
\alias{chk_unique_key}
\title{Check that a set of columns forms a unique key}
\usage{
chk_unique_key(cols)
}
\arguments{
\item{cols}{Character vector of key columns.}
}
\value{
A check function taking a data frame.
}
\description{
Check that a set of columns forms a unique key
}
