% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-frame.R
\name{is_ok}
\alias{is_ok}
\title{Did every check in a report pass?}
\usage{
is_ok(report)
}
\arguments{
\item{report}{A \code{framecheck_report}.}
}
\value{
Logical scalar.
}
\description{
Did every check in a report pass?
}
